move all registry handling under the Registry module (#2183)
