+++ test/pkg.jl
@@ -523,7 +523,7 @@ end end
 
 @testset "printing of stdlib paths, issue #605" begin
     path = Pkg.Types.stdlib_path("Test")
-    @test Pkg.Types.pathrepr(path) == "`@stdlib/Test`"
+    @test Pkg.pathrepr(path) == "`@stdlib/Test`"
 end
 
 @testset "stdlib_resolve!" begin
